+++ NEWS.md
@@ -1,5 +1,8 @@
 # RRRR 1.0.0.9000
 
+* Improved documentation
+* Bug fix: update.RRRR now includes previous history
+
 
 # RRRR 1.0.0