+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Improved documentation
 * Bug fix: update.RRRR now includes previous history
+* Improved safety: added more checks in RRRR and ORRRR
 
 
 # RRRR 1.0.0