+++ NEWS.md
@@ -3,6 +3,7 @@
 * Improved documentation
 * Bug fix: update.RRRR now includes previous history
 * Improved safety: added more checks in RRRR and ORRRR
+* plot.RRRR run time x aes bug fix.
 
 
 # RRRR 1.0.0