+++ NEWS.md
@@ -3,7 +3,8 @@
 * Improved documentation
 * Bug fix: update.RRRR now includes previous history
 * Improved safety: added more checks in RRRR and ORRRR
-* plot.RRRR run time x aes bug fix.
+* Bug fix: plot.RRRR run time x aes wrong labels are now fixed.
+* Bug fix: coef.RRR method now handles different Q and R
 
 
 # RRRR 1.0.0